Support bundles without Openstack layer
This changes the logic of is_using_bundle_builder and
is_using_automatic_placement to handle bundles without an Openstack
layer.

Note that this will return True if any single layer is using BB or AP,
so it may not work with the corner case of using BB/AP for one layer but
not another. That would require a larger architectural change that
probably is not worth investing in.